Moved subscriber registration within extension method to simplify builder interface
