+++ src/Workleap.DomainEventPropagation.Subscription.PullDelivery.Tests/Workleap.DomainEventPropagation.Subscription.PullDelivery.Tests.csproj
@@ -8,7 +8,6 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Azure.Identity" Version="1.7.0" />
     <PackageReference Include="FakeItEasy" Version="8.0.0"/>
     <PackageReference Include="GSoft.Extensions.Xunit" Version="1.0.1"/>
     <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" Version="6.0.20"/>
