+++ src/Workleap.DomainEventPropagation.Subscription.PullDelivery/IEventPropagationSubscriberBuilder.cs
@@ -5,6 +5,4 @@ namespace Workleap.DomainEventPropagation;
 public interface IEventPropagationSubscriberBuilder
 {
     IServiceCollection Services { get; }
-
-    internal void ConfigureSubscriber(Action<EventPropagationSubscriptionOptions> configure, string optionsSectionName);
 }
